多加盟商活动优化-邀请返券

+++ src/main/java/com/xiliulou/electricity/entity/JoinShareActivityHistory.java
@@ -71,7 +71,7 @@ public class JoinShareActivityHistory {
     /**
      * 加盟商Id
      */
-    private Integer franchiseeId;
+    private Long franchiseeId;
 
     //已参与
     public static Integer STATUS_INIT = 1;
