+++ src/main/java/com/xiliulou/electricity/entity/ShareActivityRecord.java
@@ -62,7 +62,7 @@ public class ShareActivityRecord {
     /**
      * 加盟商Id
      */
-    private Integer franchiseeId;
+    private Long franchiseeId;
 
     //初始化
     public static Integer STATUS_INIT = 1;
